09-mlp-numpy/mnist-serving/requirements.txt
dill
numpy
Pillow
streamlit
streamlit-drawable-canvas
tqdm
